requirements.txt
uvicorn
fastapi
pydantic
openai
python-dotenv